# shared/json_handler/json_handler.go
package json_handler

import (
	"encoding/json"
	"net/http"

	"github.com/joolshouston/pismo-technical-test/shared/model"
)

type ErrorResponse struct {
	Error string `json:"error"`
}

func WriteJSON(w http.ResponseWriter, status int, data interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if err := json.NewEncoder(w).Encode(data); err != nil {
		http.Error(w, `{"error":"failed to encode response"}`, http.StatusInternalServerError)
	}
}

func WriteError(w http.ResponseWriter, status int, errMsg string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)

	errResp := model.ErrorResponse{
		Message: errMsg,
		Status:  status,
	}

	if err := json.NewEncoder(w).Encode(errResp); err != nil {
		http.Error(w, `{"error":"failed to encode error response"}`, http.StatusInternalServerError)
	}
}